--- test/controllers/species_lists/name_lists_controller_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require("test_helper")
-
-module SpeciesLists
-  class NameListsControllerTest < FunctionalTestCase
-  end
-end
